api/requirements.txt
# Requirements for the FastAPI server and services
# python 3.11

# build
setuptools
setuptools-scm

# linting
ruff
mypy
typing_extensions
types-python-dateutil
types-pyyaml
pyupgrade

# testing
httpx
pytest
pytest-cov

# run
fastapi~=0.88.0
fastapi-mail
motor==3.1.1
pydantic==1.10.4
uvicorn[standard]
uvloop
httptools
gunicorn
beanie==1.16.6
pyyaml~=6.0
starlette==0.22
pydantic_computed
python-dateutil
psycopg2-binary
sqlalchemy
sqlalchemy-utils
firebase-admin
types-python-dateutil
types-pyyaml
python-dotenv
